Switch to long.

+++ src/codr7/shi/libraries/core/TInt.java
@@ -4,13 +4,13 @@ import codr7.shi.IType;
 import codr7.shi.IValue;
 import codr7.shi.ScriptType;
 
-public final class TInt extends ScriptType<Integer> {
+public final class TInt extends ScriptType<Long> {
     public TInt(final String name, final IType... parents) {
         super(name, parents);
     }
 
     @Override
     public boolean asBool(final IValue value) {
-        return value.cast(this) != 0;
+        return value.cast(this) != 0L;
     }
 }
